+++ lib/src/commands/do/list/organizations.dart
@@ -7,10 +7,11 @@
 import 'package:args/command_runner.dart';
 import 'package:gg_console_colors/gg_console_colors.dart';
 import 'package:gg_log/gg_log.dart';
+
 import '../../../backend/list_backend.dart';
 import '../../../backend/workspace_utils.dart';
 
-/// Command to list all organizations from repos in the master workspace.
+/// Command to list all organizations from repos in the ocean.
 class ListOrganizationsCommand extends Command<dynamic> {
   /// Constructor with optional workspace path.
   ListOrganizationsCommand({
@@ -18,7 +19,7 @@ class ListOrganizationsCommand extends Command<dynamic> {
     String? workspacePath,
     // coverage:ignore-start
   }) : workspacePath =
-            workspacePath ?? WorkspaceUtils.defaultMasterWorkspacePath();
+            workspacePath ?? WorkspaceUtils.defaultOceanWorkspacePath();
   // coverage:ignore-end
 
   /// The log function.
@@ -31,7 +32,7 @@ class ListOrganizationsCommand extends Command<dynamic> {
   String get name => 'orgs';
 
   @override
-  String get description => 'List all organizations of the master workspace';
+  String get description => 'List all organizations of the ocean';
 
   @override
   Future<void> run() async {